Switch Mocha to mostly promise based
Mocha supports promises out of the box, which is way better looking than
callback functions and allows for more concise code. Furthermore, all
tests involving the database should do a db.sync that forces an update,
this handles all the previous problems of tables not existing, foreign
keys and truncating, and all other problems. The database should be synced
only before tests, and before establishing a connection. Everything should
wait on it to finish.